Got serial Comms working #2 #9
Deleted a bunch of legacy stuff from arduino, as we are no longer using
the TFT LCD screen.

have the Arduino sending a constant message that python can pull from.
Right now, python grabs every single line that is available, and dumps
them into a list. If it is blank and the arduino isnt updating, return
None.

Created a method to request status from the arduino with a particular
call string. Still needs to be implemented arduino-side.

As of right now, Arduino comms are on COM6. Need to rewrite to
constantly repoll the ports until it finds the correct one.