worked on #9, #7, #5
**drinks.py
Instead of list accepting from drinks.py, drinks.py will instantiate a
singleton instance of a drink, which is constructed, sent off to
arduino, and then dumped upon completion of rating/feedback process.

**arduinocomm.py
now has a thread that just listens for responses, on a 3 second timer
This will eventaully set flags for other classes to act upon depending
on responses from arduino

** userinterface.py
now the dispense button sends off a constructor request to drinks.py,
it then prints out the drink for logging purposes.
Once this is created, it autogenerates a cleaned list of ingredients for
passing to the arduino(list form, from tuple)