Closed #14, worked on #18
Sliders are essentially perfect now.
Only possible improvement is to make them balance based on each other.
However this is not crucial
Changed UI and Arduinocomm to take a list of ingredient proportions as
opposed to the actual drink. This allows us to modify on the fly
pre-pour.
Removed a whole bunch of logging calls as they were superfluous.
Moved the gridding around for easier viewing.
Started work on slidersAdjusted to dynamically move them based on each
other.
PourIt now uses a list of proportions
Changed the way sliders are hidden/shown. Previously they would re-use
existing sliders and hide unused, now we hide all, and unhide/modify
ones that are necessary.