Got dispenses working with new UI worked on #20
Way it works:
IDs of the sliders are set to the ingredient names.
Upon finishing hitting dispense, it cycles through all sliders, grabs
their IDS and values, and puts them into a dict as follows { ID: Value }
and then returns it to the webserver through an ajax call. Upon
reception, the server unpacks the json, and runs it through a method
that converts it into a cleaned list for dispensing.