Closed #40
OK. lots of work here. First of all, the new database is now hooked in
and displays drinks properly.
The way it works is:

The script pulls all available drinks based on what is in the canisters.

Then, for each drink in that list, it grabs all associated ingredients,
and fills in zeros in the dictionary for non-necessary ingredients.
(this is for the template to parse it correctly)