Slightly got ingredient selection working.  #44
Still need to implement selectable ones, but it's working with hardcoded
ones.
Also I hooked it into the Upbutton, will have to remove that later.
Added 12 checkboxes, they need to get fixed and functional. Added
drinklist.tpl, which can dynamically modify the drink list without
breaking bindings(which is why we had to switch from list clicks to the
div click.
Wrote the ajax call that hits the server and reorganizes the table based
on selected ingredients.