Closed #44, closed #43
Drinks now only populate needed ingredients in their dicts, as opposed
to all 12. This is application wide now. I dumped the drinks.py file as
it was only holding the program back in terms of speed. Everything is
now dictionaries.
The Drink creation also appears to be working fully. Just have to spruce
the ol' girl up. The getDrinkProto has been reworked to go through all
the keys of the dict instead of iterating against the canister list.
This will provide speed increases as dict lookups are O(1) even if there
are one or two extras, its still faster than querying a list
comprehension